Rating CRUD : Delete

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/services/servicesImpl/CurvePointServiceImpl.java
@@ -73,7 +73,11 @@ public class CurvePointServiceImpl implements CurvePointService {
      */
     @Override
     public void delete(Integer id) {
-        curvePointRepository.deleteById(id);
-        log.debug("curve point deleted id=[{}]", id);
+        if (curvePointRepository.findById(id).isPresent()) {
+            curvePointRepository.deleteById(id);
+            log.debug("curve point deleted id=[{}]", id);
+        } else {
+            log.debug("curve point not found =[{}]", id);
+        }
     }
 }
